+++ cmd/light-client/BUILD.bazel
