+++ cmd/light-client/debug/README.md
