+++ cmd/light-client/debug/trusted-sync-comm-root/main.go
