+++ cmd/light-client/sync/verify.go
